Feat(Wiki): Jstree에서 선택된 항목을 조회 (날짜 등 부가정보 추가 조회 필요)
