+++ mos/src/main/resources/com/mos/domain/wiki/repository/WikiApiRepository.xml
@@ -43,6 +43,15 @@
 		WHERE wiki_no = #{wikiNo};
 	</select>
 
+  <!--	Wiki의 상세 내역을 반환한다.-->
+  <select id="findWikiByWikiNo" resultType="WikiDto">
+    SELECT
+      wiki_no, study_no, title, wiki.member_no, username, content_created_date, content_updated_date, content, likes
+    FROM wiki
+    JOIN member ON wiki.member_no = member.member_no
+    WHERE wiki_no = #{wikiNo};
+	</select>
+
 	<update id="updateByWikiNo" parameterType="JstreeWikiDto">
 		UPDATE wiki
 		SET
