+++ mos/src/main/resources/com/mos/domain/wiki/repository/WikiRepository.xml
@@ -7,28 +7,28 @@
 
   <select id="listByStudyNo" parameterType="int" resultType="WikiDto">
     SELECT wiki_no, title, ordr, layer, layer_ordr
-    FROM curriculum
+    FROM wiki
     WHERE study_no = #{studyNo}
-    ORDER BY ordr, layer, layer_ordr;
+    ORDER BY ordr;
   </select>
 
   <select id="listByWikiNo" parameterType="int" resultType="WikiDto">
     SELECT wiki_no, title, content, likes, hit_count
-    FROM curriculum
+    FROM wiki
     ORDER BY wiki_no DESC;
   </select>
 
   <select id="getByWikiNo" parameterType="int" resultType="WikiDto">
     SELECT
-      wiki_no, study_no, title, curriculum.member_no, username, content_created_date, content_updated_date, content, likes
-    FROM curriculum
-    JOIN member ON curriculum.member_no = member.member_no
+      wiki_no, study_no, title, wiki.member_no, username, content_created_date, content_updated_date, content, likes
+    FROM wiki
+    JOIN member ON wiki.member_no = member.member_no
     WHERE wiki_no = #{wikiNo};
   </select>
 
   <update id="updateWiki" parameterType="WikiDto">
     UPDATE
-      curriculum
+      wiki
     SET
       title=#{title},
       content=#{content}
@@ -49,20 +49,20 @@
   <delete id="deleteWiki" parameterType="int">
     DELETE
       FROM
-        curriculum
+        wiki
     WHERE
       wiki_no=#{wikiNo}
   </delete>
 
   <insert id="add" parameterType="WikiDto">
     INSERT INTO
-      curriculum(wiki_no, study_no, member_no, title, content)
+      wiki(wiki_no, study_no, member_no, title, content)
     VALUES(#{wikiNo}, #{studyNo}, 1, #{title}, #{content});
   </insert>
 
   <select id="getFirstWiki" parameterType="int" resultType="_int">
     SELECT wiki_no
-    FROM curriculum
+    FROM wiki
     WHERE study_no = #{studyNo}
     ORDER BY ordr ASC
     LIMIT 1;
