+++ mos/src/main/resources/templates/study/search.html
@@ -1,63 +1,41 @@
 <div data-th-replace="header::header">머리말</div>
 
-<div class="content-container page-margin wrapper-mos">
+<div class="container page-margin wrapper-mos">
   <h1>스터디 모집 목록</h1>
 
   <style>
-    .content-container {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start; /* 내부 요소를 왼쪽 정렬 */
-      padding: 20px; /* 컨테이너 주변 여백 */
-    }
-
     #searchForm {
-      align-self: flex-end; /* 검색창을 오른쪽에 배치 */
-      margin-bottom: 20px; /* 검색창 아래 여백 */
-    }
-
-    .search-container {
       display: flex;
-      justify-content: flex-start; /* 내부 요소를 왼쪽 정렬 */
-      align-items: center; /* 내부 요소를 수직 중앙 정렬 */
-      gap: 10px; /* 각 요소 사이의 간격 조정 */
+      justify-content: flex-end;
+      align-items: center;
     }
 
-    .search-container label {
-      white-space: nowrap; /* 라벨 텍스트가 넘치지 않도록 줄바꿈 방지 */
+    #searchForm > div {
+      margin-left: 10px; /* 각 요소 사이의 간격 조정 */
     }
 
     #type {
       width: 160px;
     }
-
-    #keyword {
-      flex: 1; /* 입력 필드가 가능한 최대 너비를 가질 수 있도록 설정 */
-    }
-
-    button[type="submit"] {
-      min-width: 80px; /* 최소 너비 설정 */
-    }
   </style>
 
   <form action="/study/search" method="get" id="searchForm">
-    <div class="search-container">
-      <div>
-        <label for="type"></label>
-        <select name="type" id="type">
-          <option value="">검색 유형 선택</option>
-          <option value="title">제목</option>
-          <option value="introduction">모집글 내용</option>
-          <option value="tag">기술스택</option>
-        </select>
-      </div>
-      <div>
-        <label for="keyword"></label>
-        <input type="text" id="keyword" name="keyword" placeholder="검색어를 입력하세요.">
-      </div>
-      <button type="submit" onclick="clearSearch()">검색</button>
+    <div>
+      <label for="type"></label>
+      <select name="type" id="type">
+        <option value="">검색 유형 선택</option>
+        <option value="title">제목</option>
+        <option value="introduction">모집글 내용</option>
+        <option value="tag">기술스택</option>
+      </select>
     </div>
+    <div>
+      <label for="keyword"></label>
+      <input type="text" id="keyword" name="keyword" placeholder="검색어를 입력하세요.">
+    </div>
+    <button type="submit">검색</button>
   </form>
+
 
   <table id="studyTable" class="table table-striped">
     <thead>
@@ -72,19 +50,7 @@
     </tr>
     </thead>
     <tbody>
-    <tr data-th-each="study : ${studyList}" data-th-object="${study}">
-      <td data-th-text="*{studyNo}">100</td>
-      <td><a data-th-href="@{view(studyNo=*{studyNo})}" data-th-text="*{title}">OO스터디 모집합니다.</a></td>
-      <td data-th-text="*{memberNo}">스터디장</td>
-      <td data-th-text="*{intake}">5</td>
-      <td data-th-text="*{method}">오프라인</td>
-      <td data-th-text="*{recruitmentDeadline}">2024-03-27</td>
-      <td>
-        <div data-th-each="tag : ${study.tagList}" data-th-object="${tag}">
-          <a data-th-text="*{name}">Java</a>
-        </div>
-      </td>
-    </tr>
+    <!-- 여기에 JavaScript로 데이터를 삽입할 예정 -->
     </tbody>
   </table>
   <div class="text-center">
@@ -93,6 +59,7 @@
     </div>
   </div>
   <div id="sentinel"></div>
+
 
 </div>
 <div data-th-replace="footer::footer">꼬리말</div>
