+++ mos/src/main/resources/templates/wiki/jstreeTest.html
@@ -11,6 +11,6 @@
 
   <script src="/plugins/jstree/jstree.js"></script>
   <link rel="stylesheet" href="/plugins/jstree/themes/default/style.css">
-  <script src="/js/wiki/jstreeTest.js"></script>
+  <script src="/js/wiki/wikitree-edit.js"></script>
 </div>
 <div th:replace="footer::footer">꼬리말</div>
